refactor: de-spine the Mavros prover so prover is field-generic
The prover crate still carried a bn254-welded tail: the MavrosR1CSProver
trait and its impl for WhirR1CSScheme<Bn254Field>, which named Bn254Field
and depended on mavros-vm/mavros-artifacts. This blocked relocating
Bn254Field out of common (prover depends on common only, and backend
already depends on prover, so prover referencing a backend Bn254Field
would close a cycle).

Expose the generic proving primitives the Mavros prover composes
(prove_from_alphas is now pub; run_zk_sumcheck_prover and WhirR1CSCommitment
re-exported), and move MavrosR1CSProver into provekit-backend-bn254 as
mavros_prove. The prover crate drops its mavros dependencies and its
Bn254Field/FieldElement references — its direct deps are now just ark-ff,
ark-std, whir, and common. The logic is unchanged; this is a pure
relocation, not a genericization of the Mavros path (that still waits on a
field-agnostic Mavros VM upstream).

72/72 e2e green.
