+++ provekit/prover/src/lib.rs
@@ -1,5 +1,5 @@
 mod whir_r1cs;
 
-#[cfg(not(target_arch = "wasm32"))]
-pub use whir_r1cs::MavrosR1CSProver;
-pub use whir_r1cs::WhirR1CSProver;
+pub use whir_r1cs::{
+    prove_from_alphas, run_zk_sumcheck_prover, WhirR1CSCommitment, WhirR1CSProver,
+};
